+++ packages/pages/src/app/UI/SpeechRecognitionSimpleEvents.jsx
@@ -60,7 +60,7 @@ const SpeechRecognitionSimpleEvents = () => {
               </React.Fragment>
             ) : event.type === 'cognitiveservices' ? (
               <span className="badge badge-light">
-                {event.type}:{event.data.type}
+                {event.type}:{event.data?.type}
               </span>
             ) : (
               <span className="badge badge-secondary">{event.type}</span>