+++ packages/web-speech-cognitive-services/src/SpeechServices/SpeechToText/SpeechRecognitionResultList.ts
@@ -1,0 +1,8 @@
+import FakeArray from './FakeArray';
+import type SpeechRecognitionResult from './SpeechRecognitionResult';
+
+export default class SpeechRecognitionResultList extends FakeArray<SpeechRecognitionResult> {
+  constructor(result: readonly SpeechRecognitionResult[]) {
+    super(result);
+  }
+}